add some modifications

+++ lib/feature/countries/domain/entity/country_entity.dart
@@ -3,7 +3,7 @@ class CountryEntity {
   final String flag;
   final String region;
   final String subRegion;
-  final double population;
+  final int population;
   final double area;
   final List<String> timeZones;
